pyproject.toml
[tool.poetry]
name = "ene"
version = "0.1.1"
description = "Automatically track and sync anime watching progress"
authors = ["Peijun Ma <peijun.ma@protonmail.com>","Justin Sedge <sedgejustin@gmail.com>"]
license = "GPL-3.0+"
readme = "README.md"
repository = "https://github.com/MaT1g3R/ene"
homepage = "https://github.com/MaT1g3R/ene"
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
]

[tool.poetry.dependencies]
python = "^3.6"
toml = "*"
requests = "*"
python-mpv = "*"
python-vlc = "*"
fuzzywuzzy = "*"
PySide2 = "*"
python-Levenshtein = "*"
attrs = "*"
option = "*"
peewee = "^3.8"
pyinstaller = "^3.4"

[tool.poetry.dev-dependencies]
pytest = "*"
flake8 = "*"
pylint = "*"
pytest-cov = "*"
responses = "*"
bumpversion = "^0.5.3"
faker = "^1.0"
